requirements.txt
gspread==6.2.0
oauth2client==4.1.3
python-dotenv==1.1.0
twilio==9.5.2
